+++ src/renderer/Root.tsx
@@ -16,7 +16,6 @@ import ErrorView from "./views/error/ErrorView";
 import { InputPrivateKeyView } from "./views/account/reset/InputPrivateKeyView";
 import { RegisterPrivateKeyView } from "./views/account/reset/RegisterPrivateKeyView";
 import PreloadView from "./views/lobby/PreloadView";
-import ActivationAccountView from "./views/lobby/ActivationAccountView";
 
 class Root extends React.Component {
   render() {
@@ -29,11 +28,6 @@ class Root extends React.Component {
           <Route exact path="/login/mining" component={MiningView} />
           <Route exact path="/lobby" component={LobbyView} />
           <Route exact path="/lobby/preload" component={PreloadView} />
-          <Route
-            exact
-            path="/lobby/activation"
-            component={ActivationAccountView}
-          />
           <Route exact path="/error" component={ErrorView} />
           <Route exact path="/account" component={AccountView} />
           <Route exact path="/account/create" component={CreateAccountView} />
